+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### 12.15.0
 
 * `[spinner]` Added new package for showing spinner animation for long-running tasks
+* `[timeutil]` Added high precision mode for `ShortDuration`
 
 ### 12.14.1